- Get rid of action endpoint (will move to OPTIONS) - Make Action an inner class of Role
