+++ dockstore-webservice/src/main/java/io/dockstore/webservice/permissions/Role.java
@@ -20,4 +20,22 @@ public enum Role {
     OWNER,
     WRITER,
     READER;
+
+    public enum Action {
+        WRITE("write"),
+        READ("read"),
+        DELETE("delete"),
+        SHARE("share");
+
+        private final String name;
+
+        Action(String name) {
+            this.name = name;
+        }
+
+        @Override
+        public String toString() {
+            return this.name;
+        }
+    }
 }
